refactor: refactored episodes screen.
