--- ui/src/main/java/com/cairosquad/ui/details/EpisodesScreen.kt
@@ -1,449 +1,0 @@
-package com.cairosquad.ui.details
-
-import android.os.Build
-import android.widget.Toast
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.navigationBars
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.statusBars
-import androidx.compose.foundation.layout.systemBars
-import androidx.compose.foundation.layout.windowInsetsPadding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicText
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.BlurredEdgeTreatment
-import androidx.compose.ui.draw.blur
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush.Companion.verticalGradient
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.lerp
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.cairosquad.design_system.R
-import com.cairosquad.design_system.basic_component.AppBar
-import com.cairosquad.design_system.basic_component.AppDropDownMenu
-import com.cairosquad.design_system.basic_component.Icon
-import com.cairosquad.design_system.theme.Theme
-import com.cairosquad.safe_image_viewer.safe_image_viewer.SafeImageViewer
-import com.cairosquad.ui.BuildConfig
-import com.cairosquad.ui.movio_component.LoadingMovieImage
-import com.cairosquad.ui.navigation.LocalNavController
-import com.cairosquad.ui.utils.ObserveAsEffect
-import com.cairosquad.ui.utils.errorStatusToMessageResource
-import com.cairosquad.viewmodel.details.episodes.EpisodesDetailEffect
-import com.cairosquad.viewmodel.details.episodes.EpisodesDetailsInteractionListener
-import com.cairosquad.viewmodel.details.episodes.EpisodesDetailsScreenState
-import com.cairosquad.viewmodel.details.episodes.EpisodesDetailsViewModel
-
-@Composable
-fun EpisodesScreen(
-    seriesId: Long,
-    seasonNumber: Int,
-) {
-    val viewModel: EpisodesDetailsViewModel =
-        hiltViewModel<EpisodesDetailsViewModel, EpisodesDetailsViewModel.Factory> { factory ->
-            factory.create(seriesId, seasonNumber)
-        }
-    val navController = LocalNavController.current
-    val context = LocalContext.current
-    val uiState by viewModel.screenState.collectAsStateWithLifecycle()
-
-    ObserveAsEffect(viewModel.effect) { effect ->
-        when (effect) {
-            EpisodesDetailEffect.NavigateBack -> {
-                navController.popBackStack()
-            }
-
-            is EpisodesDetailEffect.ShowToast -> {
-                Toast.makeText(
-                    context,
-                    context.getString(errorStatusToMessageResource(effect.message)),
-                    Toast.LENGTH_LONG
-                ).show()
-            }
-
-            EpisodesDetailEffect.PlayEpisode -> {}
-        }
-    }
-    EpisodesScreenContent(uiState = uiState, listener = viewModel, seriesId)
-}
-
-
-@Composable
-private fun EpisodesScreenContent(
-    uiState: EpisodesDetailsScreenState,
-    listener: EpisodesDetailsInteractionListener,
-    seriesId: Long,
-    modifier: Modifier = Modifier,
-
-    ) {
-    val seasonOptions = uiState.seasons.map {
-        stringResource(
-            com.cairosquad.ui.R.string.season,
-            it.seasonNumber
-        )
-    }
-
-    val listState = rememberScrollState()
-    val density = LocalDensity.current
-    val scrollThresholdPx = with(density) { 275.dp.toPx() }
-    val progress = (listState.value / scrollThresholdPx).coerceIn(0f, 1f)
-
-    val animatedStartColor = lerp(
-        start = Theme.color.surfaces.statusBarShadow,
-        stop = Theme.color.surfaces.surface,
-        fraction = progress
-    )
-    val animatedEndColor = lerp(
-        start = Color.Transparent,
-        stop = Theme.color.surfaces.surface,
-        fraction = progress
-    )
-    val animatedBrush = verticalGradient(
-        colors = listOf(animatedStartColor, animatedEndColor)
-    )
-    when (uiState.episodesSectionState) {
-        EpisodesDetailsScreenState.ScreenStatus.ERROR -> {
-            DetailsFailContent(onTryAgainClick = listener::onRefresh)
-
-        }
-    else->{
-        Box(
-            modifier = modifier
-                .fillMaxSize()
-                .background(Theme.color.surfaces.surface)
-                .windowInsetsPadding(WindowInsets.navigationBars)
-                .verticalScroll(listState)
-        ) {
-            Box {
-                when (uiState.basicDetailsSectionState) {
-                    EpisodesDetailsScreenState.ScreenStatus.LOADING -> {}
-                    EpisodesDetailsScreenState.ScreenStatus.SUCCESS -> {
-                        if (uiState.season.posterUrl.isNotEmpty()) {
-                            Box {
-                                SafeImageViewer(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .height(400.dp)
-                                        .then(
-                                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-                                                Modifier.blur(
-                                                    16.dp,
-                                                    edgeTreatment = BlurredEdgeTreatment.Unbounded
-                                                )
-                                            } else {
-                                                Modifier
-                                            }
-                                        )
-                                        .offset(y = (-28).dp),
-                                    model = BuildConfig.IMAGE_BASE_URL + uiState.season.posterUrl,
-                                    contentDescription = "",
-                                    blur = if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) 16 else 0,
-                                    isBlurForced = true
-                                )
-                                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
-                                    Box(
-                                        modifier = Modifier
-                                            .fillMaxWidth()
-                                            .height(60.dp)
-                                            .align(Alignment.BottomCenter)
-                                            .background(
-                                                brush = verticalGradient(
-                                                    colors = listOf(
-                                                        Theme.color.surfaces.surface.copy(alpha = 0.00f),
-                                                        Theme.color.surfaces.surface.copy(alpha = 0.10f),
-                                                        Theme.color.surfaces.surface.copy(alpha = 0.50f),
-                                                        Theme.color.surfaces.surface.copy(alpha = 0.90f),
-                                                        Theme.color.surfaces.surface,
-                                                    )
-                                                )
-                                            )
-                                    )
-                                }
-
-                            }
-                        }
-                    }
-
-                    EpisodesDetailsScreenState.ScreenStatus.ERROR -> {}
-                }
-            }
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .windowInsetsPadding(WindowInsets.systemBars)
-                    .heightIn(max = 10000.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                userScrollEnabled = false,
-            ) {
-                item {
-                    when (uiState.basicDetailsSectionState) {
-                        EpisodesDetailsScreenState.ScreenStatus.LOADING -> {
-                            LoadingMovieImage(
-                                Modifier
-                                    .padding(top = 56.dp, bottom = 24.dp)
-                                    .size(height = 260.dp, width = 200.dp)
-                            )
-                        }
-
-                        EpisodesDetailsScreenState.ScreenStatus.SUCCESS -> {
-                            if (uiState.season.posterUrl.isNotEmpty()) {
-                                SafeImageViewer(
-                                    modifier = Modifier
-                                        .padding(top = 56.dp, bottom = 24.dp)
-                                        .size(height = 260.dp, width = 200.dp)
-                                        .clip(RoundedCornerShape(8.dp)),
-                                    model = BuildConfig.IMAGE_BASE_URL + uiState.season.posterUrl,
-                                    contentDescription = "",
-                                    loadingPlaceholder = {
-                                        LoadingMovieImage(
-                                            Modifier.size(height = 260.dp, width = 200.dp)
-                                        )
-                                    }
-                                )
-                            } else {
-                                Box(
-                                    modifier = Modifier
-                                        .padding(top = 56.dp, bottom = 24.dp)
-                                        .size(height = 260.dp, width = 200.dp)
-                                        .clip(RoundedCornerShape(8.dp))
-                                        .background(Theme.color.system.defaultImageBackground),
-                                    contentAlignment = Alignment.Center
-                                ) {
-                                    Icon(
-                                        modifier = Modifier.size(24.dp),
-                                        imageVector = ImageVector.vectorResource(id = R.drawable.image_icon),
-                                        contentDescription = stringResource(R.string.default_image_icon),
-                                        tint = Color(0xFFEFF1F5)
-                                    )
-                                }
-                            }
-                        }
-
-                        EpisodesDetailsScreenState.ScreenStatus.ERROR -> {}
-                    }
-                }
-                item {
-                    when (uiState.basicDetailsSectionState) {
-                        EpisodesDetailsScreenState.ScreenStatus.LOADING -> {
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(horizontal = 16.dp, vertical = 16.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                LoadingMovieImage(
-                                    modifier = Modifier
-                                        .clip(CircleShape)
-                                        .size(height = 26.dp, width = 32.dp)
-                                )
-                                Spacer(modifier = Modifier.weight(1f))
-                                LoadingMovieImage(
-                                    modifier = Modifier
-                                        .clip(CircleShape)
-                                        .size(height = 26.dp, width = 96.dp)
-                                )
-                            }
-                        }
-
-                        EpisodesDetailsScreenState.ScreenStatus.SUCCESS -> {
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(horizontal = 16.dp, vertical = 16.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                BasicText(
-                                    modifier = Modifier.padding(bottom = 8.dp),
-                                    text = stringResource(
-                                        com.cairosquad.ui.R.string.episodes,
-                                        uiState.season.episodesCount
-                                    ),
-                                    style = Theme.textStyle.headline.mediumMedium18.copy(
-                                        color = Theme.color.surfaces.onSurface
-                                    )
-                                )
-                                Spacer(modifier = Modifier.weight(1f))
-                                AppDropDownMenu(
-                                    selectedText = stringResource(
-                                        R.string.season,
-                                        uiState.selectedSeasonNumber
-                                    ),
-                                    options = seasonOptions,
-                                    imgRes = R.drawable.drop_down_arrow,
-                                    onOptionSelected = { selectedSeason ->
-                                        val seasonNum =
-                                            seasonOptions.indexOfFirst {
-                                                selectedSeason == it
-                                            }.takeIf { it != -1 } ?: 1
-                                        listener.onSeasonSelected(seriesId, seasonNum)
-                                    },
-                                )
-                            }
-                        }
-
-                        EpisodesDetailsScreenState.ScreenStatus.ERROR -> {}
-                    }
-                }
-                when (uiState.episodesSectionState) {
-                    EpisodesDetailsScreenState.ScreenStatus.LOADING -> {
-                        items(10) {
-                            LoadingMovieImage(
-                                modifier = Modifier
-                                    .padding(horizontal = 16.dp, vertical = 12.dp)
-                                    .height(74.dp)
-                                    .fillMaxSize(),
-                            )
-                        }
-                    }
-
-                    EpisodesDetailsScreenState.ScreenStatus.SUCCESS -> {
-                        items(uiState.episodes) { episode ->
-                            EpisodeCard(
-                                episodeNumber = episode.number.toString().padStart(2, '0'),
-                                episodeImageUrl = BuildConfig.IMAGE_BASE_URL + episode.imageUrl,
-                                episodeName = episode.name,
-                                episodeDuration = episode.runtime,
-                                episodeRating = String.format("%.1f", episode.rating),
-                            )
-                        }
-                    }
-
-                    EpisodesDetailsScreenState.ScreenStatus.ERROR -> {}
-                }
-            }
-        }}
-    }
-    AppBar(
-        modifier = Modifier
-            .background(brush = animatedBrush)
-            .windowInsetsPadding(WindowInsets.statusBars)
-            .fillMaxWidth(),
-        onBackButtonClicked = listener::onBackClick,
-    )
-}
-
-@Composable
-fun EpisodeCard(
-    modifier: Modifier = Modifier,
-    episodeName: String,
-    episodeNumber: String,
-    episodeDuration: Int,
-    episodeRating: String,
-    episodeImageUrl: String? = null
-) {
-    Row(
-        modifier = modifier
-            .padding(bottom = 12.dp)
-            .padding(horizontal = 16.dp)
-            .background(Theme.color.surfaces.surface)
-            .fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Box(
-
-        ) {
-            SafeImageViewer(
-                contentDescription = "Episode Image",
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .size(width = 100.dp, height = 74.dp)
-                    .clip(RoundedCornerShape(8.dp)),
-                model = episodeImageUrl.toString()
-            )
-            Box(
-                modifier = Modifier
-                    .size(20.dp)
-                    .clip(CircleShape)
-                    .background(Color.White.copy(alpha = 0.4f))
-                    .blur(16.dp)
-                    .align(Alignment.Center)
-            )
-            Image(
-                painter = painterResource(R.drawable.outline_play),
-                contentDescription = "Play",
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .size(width = 7.dp, height = 8.dp)
-            )
-        }
-        Column(
-            modifier = Modifier.padding(end = 8.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(4.dp)
-            ) {
-                BasicText(
-                    text = episodeName,
-                    style = Theme.textStyle.title.mediumMedium14.copy(color = Theme.color.surfaces.onSurface),
-                    modifier = Modifier.weight(1f)
-                )
-                Image(
-                    painter = painterResource(R.drawable.review_star),
-                    contentDescription = "Rating",
-                    modifier = Modifier.size(16.dp)
-                )
-                BasicText(
-                    text = episodeRating,
-                    style = Theme.textStyle.label.smallRegular12.copy(color = Theme.color.surfaces.onSurface)
-                )
-            }
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(4.dp)
-            ) {
-                BasicText(
-                    text = stringResource(com.cairosquad.ui.R.string.episode, episodeNumber),
-                    style = Theme.textStyle.label.smallRegular12.copy(
-                        color = Theme.color.surfaces.onSurfaceContainer,
-                    )
-                )
-                BasicText(
-                    text = " • ",
-                    style = Theme.textStyle.label.smallRegular12.copy(
-                        color = Theme.color.surfaces.onSurfaceContainer,
-                    )
-                )
-                BasicText(
-                    text = stringResource(com.cairosquad.ui.R.string.m, episodeDuration),
-                    style = Theme.textStyle.label.smallRegular12.copy(
-                        color = Theme.color.surfaces.onSurfaceContainer,
-                    )
-                )
-            }
-        }
-    }
-}
